Docs: Some changes Client code
