+++ client/pages/Admin/Doctorlist.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-
+import Layout from "../../src/componnents/Layout";
 import axios from "axios";
 import { showLoading, hideLoading } from "../../src/redux/alertSlice";
 import { useDispatch } from "react-redux";
@@ -82,37 +82,41 @@ export const Doctorlist = () => {
     },
   ];
   return (
-    <div className="p-4">
-      <p className="text-3xl font-bold font-serif">Doctors List</p>
-      <table className="min-w-full divide-y divide-gray-200 shadow-lg mt-6 ">
-        <thead className="bg-gray-300">
-          <tr>
-            {doctorcolumns.map((doctor) => (
-              <th
-                key={doctor.id}
-                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider rounded-xl"
-              >
-                {doctor.title}
-              </th>
-            ))}
-          </tr>
-        </thead>
-        <tbody className="bg-white divide-y divide-gray-200">
-          {doctors.map((doctord, doctorIndex) => (
-            <tr key={doctorIndex}>
-              {doctorcolumns.map((column) => (
-                <td
-                  key={`${doctord.index}-${column.id}`}
-                  className="px-6 py-4 whitespace-nowrap text-sm text-gray-500 rounded-xl"
-                >
-                  {doctord[column.id]}
-                  {column.id === "action" ? column.render("", doctord) : ""}
-                </td>
+    <>
+      <Layout>
+        <div className="p-4">
+          <p className="text-3xl font-bold font-serif">Doctors List</p>
+          <table className="min-w-full divide-y divide-gray-200 shadow-lg mt-6 ">
+            <thead className="bg-gray-300">
+              <tr>
+                {doctorcolumns.map((doctor) => (
+                  <th
+                    key={doctor.id}
+                    className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider rounded-xl"
+                  >
+                    {doctor.title}
+                  </th>
+                ))}
+              </tr>
+            </thead>
+            <tbody className="bg-white divide-y divide-gray-200">
+              {doctors.map((doctord, doctorIndex) => (
+                <tr key={doctorIndex}>
+                  {doctorcolumns.map((column) => (
+                    <td
+                      key={`${doctord.index}-${column.id}`}
+                      className="px-6 py-4 whitespace-nowrap text-sm text-gray-500 rounded-xl"
+                    >
+                      {doctord[column.id]}
+                      {column.id === "action" ? column.render("", doctord) : ""}
+                    </td>
+                  ))}
+                </tr>
               ))}
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
+            </tbody>
+          </table>
+        </div>
+      </Layout>
+    </>
   );
 };
