+++ client/pages/Applydoctor.jsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-
+import Layout from "../src/componnents/Layout";
 import { showLoading, hideLoading } from "../src/redux/alertSlice";
 import { useDispatch, useSelector } from "react-redux";
 import { toast } from "react-hot-toast";
@@ -58,6 +58,8 @@ export const Applydoctor = () => {
     }
   };
   return (
+    <>
+    <Layout>
     <div className="p-4">
       <p className="text-3xl font-bold font-serif">Apply Doctors</p>
       <hr className="border-black"></hr>
@@ -193,5 +195,7 @@ export const Applydoctor = () => {
         </button>
       </form>
     </div>
+    </Layout>
+    </>
   );
 };
