+++ client/pages/Doctor/DoctorAppointment.jsx
@@ -1,8 +1,8 @@
-
 import { useEffect, useState } from "react";
 import axios from "axios";
 import moment from "moment";
 import { toast } from "react-hot-toast";
+import Layout from "../../src/componnents/Layout";
 const DoctorAppiontments = () => {
   const [appointments, setAppointments] = useState([]);
   const columns = [
@@ -86,39 +86,47 @@ const DoctorAppiontments = () => {
     }
   };
   return (
-    <div className="p-4">
-      <p className="text-3xl font-bold font-serif">Doctor Appointments List</p>
+    <>
+      <Layout>
+        <div className="p-4">
+          <p className="text-3xl font-bold font-serif">
+            Doctor Appointments List
+          </p>
 
-      <table className="min-w-full divide-y divide-gray-200 rounded-xl mt-6 shadow-lg">
-        <thead className="bg-gray-300">
-          <tr>
-            {columns.map((list) => (
-              <th
-                key={list.id}
-                className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider rounded-xl"
-              >
-                {list.title}
-              </th>
-            ))}
-          </tr>
-        </thead>
-        <tbody className="bg-white divide-y divide-gray-200 ">
-          {appointments.map((appointment, appointmentIndex) => (
-            <tr key={appointmentIndex}>
-              {columns.map((column) => (
-                <td
-                  key={`${column.id}-${appointmentIndex}`}
-                  className="px-6 py-4 whitespace-nowrap text-sm text-gray-500 rounded-xl"
-                >
-                  {appointment[column.id]}
-                  {column.id === "action" ? column.render("", appointment) : ""}
-                </td>
+          <table className="min-w-full divide-y divide-gray-200 rounded-xl mt-6 shadow-lg">
+            <thead className="bg-gray-300">
+              <tr>
+                {columns.map((list) => (
+                  <th
+                    key={list.id}
+                    className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider rounded-xl"
+                  >
+                    {list.title}
+                  </th>
+                ))}
+              </tr>
+            </thead>
+            <tbody className="bg-white divide-y divide-gray-200 ">
+              {appointments.map((appointment, appointmentIndex) => (
+                <tr key={appointmentIndex}>
+                  {columns.map((column) => (
+                    <td
+                      key={`${column.id}-${appointmentIndex}`}
+                      className="px-6 py-4 whitespace-nowrap text-sm text-gray-500 rounded-xl"
+                    >
+                      {appointment[column.id]}
+                      {column.id === "action"
+                        ? column.render("", appointment)
+                        : ""}
+                    </td>
+                  ))}
+                </tr>
               ))}
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
+            </tbody>
+          </table>
+        </div>
+      </Layout>
+    </>
   );
 };
